Sample the configuration.
