+++ .gitignore
@@ -1,10 +1,10 @@
 *.swp
 .idea/
-api/dbconnector.php
 assets/downloads/
 bbs/attachment/
 bbs/images/
 bbsimg/
+config.php
 documents/
 dump/
 log/
