+++ api/jiekouapi.php
@@ -1,5 +1,5 @@
 <?php
-    require_once 'dbconnector.php';
+    require_once '../lib.php';
 
     if(false&&@$_REQUEST['debug']=='yes'){
         echo("GET:<br>");
