+++ bbs/sign/test.php
@@ -1,5 +1,5 @@
 <?php
-	require_once '../../assets/api/dbconnector.php';
+	require_once '../../lib.php';
 
 	header("Content-type:text/html;charset=utf-8");
 	date_default_timezone_set('Asia/Shanghai');
