+++ config.sample.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * Configuration for CAPUBBS.
+ *
+ * You need to copy this file to `config.php` to make it work.
+ *
+ * This file contains the following configrations:
+ *
+ * - MySQL settings.
+ *
+ * Reference:
+ * - https://github.com/WordPress/WordPress/blob/master/wp-config-sample.php
+ */
+
+//** MySQL settings. **//
+/** The database username. */
+define('DB_USERNAME', 'database_username_here');
+
+/** The database password. */
+define('DB_PASSWORD', 'database_password_here');
+
+/** The database hostname. */
+define('DB_HOSTNAME', 'localhost');
