+++ index/borrow.php
@@ -7,8 +7,7 @@
 <script src="/assets/js/bootstrap.min.js"></script>
 <script src="/assets/js/self-borrow.js"></script>
 <?php
-	require_once '../assets/api/dbconnector.php';
-	require_once '../assets/api/checkuser.php';
+	require_once '../lib.php';
 	$res=checkuser();
 	$username=$res[0];$rights=$res[1]; 
 	date_default_timezone_set('Asia/Shanghai');
