+++ index/transaction.php
@@ -8,7 +8,7 @@
 <body>
 
 <?php
-	require_once '../assets/api/dbconnector.php';
+	require_once '../lib.php';
 	dbconnect();
 	$statement="select number from capubbs.borrow where type=1 && state=0";
 	$results=mysql_query($statement);
